requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.8.0    # vector search (embeddings/build_embeddings.py)
fastapi==0.111.0    # only if you’ll build interface/api.py
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-text-splitters==0.3.8
langsmith==0.3.42
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
neo4j==5.28.1
networkx==3.4.2
numpy==2.2.5
openai==1.77.0
ollama-python==0.1.2
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn[standard]==0.29.0
yarl==1.20.0
zstandard==0.23.0
